Move out basis set to a separate file to avoid circular imports
